Extract out User Manager Loggable extension for file length
